refactor message to logs
